BETA 2.9 b1
The bot will now automatically start after 7.5 seconds when an update is available, disableautoupdate is true and the user isn't responding. Fixed a bug that crashed the bot when trying to comment aftera failed private profile check (#57). Added shared_secret support. Added small check that aborts the bot's startup if the user is missing the steam-user or steamcommunity npm package.