Reverted bbc08ce
I was unable to fix the comment loop activecommentprocess not defined error after two months so I decided to revert to the old loop design. I only tested this briefly, it doesn't support retryComments yet and I still need to apply changes from the following commits to the old design.